Update README and remove sample app
